Fix and align examples
