+++ examples/rcat/c/rcat.c
@@ -40,7 +40,7 @@ int send_input_data(void *carry_data) {
                 if (read_len > 0) {
                     rasta_send(data->rc, data->connection, buf, read_len);
                 }
-                // TODO: Disconnect
+
                 rasta_disconnect(data->connection);
                 return 1;
             }
